Simple handling of jobs

+++ src/Worker/Worker.php
@@ -20,13 +20,20 @@ class Worker
      */
     private $memoryLimit = 100000;
 
+    /**
+     * Worker constructor.
+     * @param \Pheanstalk\Pheanstalk $pheanstalk
+     */
     public function __construct(Pheanstalk $pheanstalk)
     {
         $this->pheanstalk = $pheanstalk;
         $this->pheanstalk->watch('jobs');
-        $this->pheanstalk->watch('commands');
     }
 
+    /**
+     * Start Worker
+     * @param null $id
+     */
     public function start($id = null)
     {
         if (!is_string($id) || strlen($id) < 5) {
@@ -37,19 +44,43 @@ class Worker
         $this->listen();
     }
 
+    /**
+     * Generate a Unique Worker Id
+     * @return mixed
+     */
     private function generateId()
     {
         return uniqid();
     }
 
+    /**
+     * Listen for jobs
+     */
     private function listen()
     {
         echo 'Worker ' . $this->id . ' starting...' . PHP_EOL;
         while ($job = $this->pheanstalk->reserve()) {
-            echo($job->getData());
-            echo PHP_EOL;
-            $this->pheanstalk->delete($job);
+            $decodedJob = json_decode($job->getData(), true);
+            if (!is_array($decodedJob) || !$this->handleJob($decodedJob)) {
+                $this->pheanstalk->bury($job);
+            } else {
+                $this->pheanstalk->delete($job);
+            }
         }
+    }
+
+    /**
+     * @param array $decodedJob
+     * @return bool
+     */
+    private function handleJob(array $decodedJob)
+    {
+        if ($decodedJob['type'] === 'order') {
+            // do whatever you want to handle an order...
+            $result = Order::handle($decodedJob);
+            return $result;
+        }
+        return false;
     }
 
 }
